Preliminary flags support.

+++ ebuild/builtin_init.bash
@@ -34,6 +34,9 @@ builtin_init()
     done
 
     if [[ -e "${PALUDIS_TMPDIR}/${CATEGORY}/${PF}" ]] ; then
+        if type -p chflags &>/dev/null; then
+            chflags -R 0 "${PALUDIS_TMPDIR}/${CATEGORY}/${PF}" || die "Couldn't remove flags from workdir"
+        fi
         rm -fr "${PALUDIS_TMPDIR}/${CATEGORY}/${PF}" || die "Couldn't remove previous work"
     fi
