+++ ebuild/builtin_tidyup.bash
@@ -20,6 +20,10 @@
 builtin_tidyup()
 {
     if [[ -e "${PALUDIS_TMPDIR}/${CATEGORY}/${PF}" ]] ; then
+        if type -p chflags &>/dev/null; then
+            echo chflags -R 0 "${PALUDIS_TMPDIR}/${CATEGORY}/${PF}" 1>&2
+            chflags -R 0 "${PALUDIS_TMPDIR}/${CATEGORY}/${PF}" || die "Couldn't remove flags from workdir"
+        fi
         echo rm -fr "${PALUDIS_TMPDIR}/${CATEGORY}/${PF}" 1>&2
         rm -fr "${PALUDIS_TMPDIR}/${CATEGORY}/${PF}" || die "Couldn't remove work"
     fi
